requirements.txt
cairosvg
pillow
requests